Dev: grammar table finalization
